fixed facebook issue

+++ src/main/java/com/gracie/barra/auth/AuthEnrichFilter.java
@@ -23,6 +23,7 @@ import com.gracie.barra.auth.dao.SessionDao;
 import com.restfb.DefaultFacebookClient;
 import com.restfb.FacebookClient;
 import com.restfb.Version;
+import com.restfb.exception.FacebookOAuthException;
 
 public class AuthEnrichFilter implements Filter {
 	private static final Logger log = Logger.getLogger(AuthEnrichFilter.class.getName());
@@ -50,16 +51,22 @@ public class AuthEnrichFilter implements Filter {
 			String accessToken = null;
 			// log.info("Session : " + req.getSession().getId());
 			accessToken = sessionDao.getToken(req.getSession().getId());
-			FacebookClient facebookClient = new DefaultFacebookClient(accessToken, Version.VERSION_2_9);
+			FacebookClient facebookClient = new DefaultFacebookClient(accessToken, Version.LATEST);
 			try {
 				if (accessToken != null && !facebookClient.debugToken(accessToken).isValid()) {
 					accessToken = null;
 					sessionDao.deleteToken(req.getSession().getId());
 				}
 			} catch (Throwable t) {
-				log.severe("Issue with token" + t.getMessage());
-				accessToken = null;
-				sessionDao.deleteToken(req.getSession().getId());
+				// Try not to care about #100
+				if (t instanceof FacebookOAuthException && ((FacebookOAuthException) t).getErrorCode() == 100) {
+					log.info("Don't care about this error " + accessToken + " " + t.getMessage());
+
+				} else {
+					log.severe("Issue with token " + accessToken + " " + t.getMessage());
+					accessToken = null;
+					sessionDao.deleteToken(req.getSession().getId());
+				}
 			}
 			if (accessToken == null) {
 				String accessCode = null;
@@ -77,25 +84,28 @@ public class AuthEnrichFilter implements Filter {
 				}
 				if (accessCode != null) {
 					try {
-						facebookClient = new DefaultFacebookClient(Version.VERSION_2_9);
+						facebookClient = new DefaultFacebookClient(Version.LATEST);
 						accessToken = facebookClient
 								.obtainUserAccessToken("123435078353209",
 										servletReq.getServletContext().getInitParameter("facebook.secret"), "", accessCode)
 								.getAccessToken();
 						sessionDao.storeToken(req.getSession().getId(), accessToken);
+						log.info("Token issued : " + accessToken);
 					} catch (Throwable t) {
-						log.severe(t.getMessage());
+						log.severe("Issue with token " + accessToken + " code " + accessCode + "" + t.getMessage());
 						sessionDao.deleteToken(req.getSession().getId());
 					}
 				}
 			}
 
 			if (accessToken != null) {
-				facebookClient = new DefaultFacebookClient(accessToken, Version.VERSION_2_9);
+				facebookClient = new DefaultFacebookClient(accessToken, Version.LATEST);
 				com.restfb.types.User user = facebookClient.fetchObject("me", com.restfb.types.User.class);
 				req.getSession().setAttribute("userEmail", user.getName());
 				req.getSession().setAttribute("userId", user.getId());
 				req.getSession().setAttribute("loginFrom", "Facebook");
+				log.info("Token " + accessToken + " belongs to  " + user.getName());
+
 			} else {
 				// remove user from session
 				req.getSession().removeAttribute("userEmail");
